Problem 2.2
Create an array of integers, array[MAX], and fill it with MAX no of
random numbers. Let then the
program print out the following:
The value of the label array (address) is: xxxxxxxxxx
First integer in the array is (array[0]) : xxxxxxxxxx
The size of an integer (number of bytes) is : xxxxxxxxx
The size of the whole array is : xxxxxxxxx
End the program by printing out the array, then double all values in
the array by use of a pointer and
print it out again.