Assignment problem 2.5
You should develop a program that calculates some statistical values
for an array of integers. Among
other things, the program will plot a histogram for the frequency of
the different input numbers in
the array.
To test it you need first to create an array of integers (table [MAX])
with MAX number of random
numbers between 0 and MAXNUMBER. Then you should write a function that
for each possible
number between 0 – MAXNUMBER calculates how many times the number
exists in the array. The
result is then stored in a new array (frequency []).
Finally, you write a function that given the array frequency [] draw a
histogram as followed example:
You should use the function and its function declaration below.
Given an array table[]={ 1,2,12, 5,1,0,0,5,9,12, 0,2,3,0} thhe program
will printout:
0 xxxx
1 xx
2 xx
3 x
5 xx
9 x
12 xx