Fixed some errors
Create a program that checks an entered Swedish person number. The
number is entered in the form
of : 7107254786 (yymmddxxxc). The number should be read in as a string
and converted to integers
for year, month , day and number.
The last digit (here 6) is a control digit and is calculated from the
other digits by an algorithm that you
can find on the internet.
The user inputs the number, the program first checks that the number of
month and day is in the
right range and after that calculate and checks the control digit. The
program then prints out the
result and asks for a new person number. This is repeated until the
user inputs a ‘q’.
The program should at least consist of the functions: main() ,
readPersnr( char *person), int
controlDigit( const char * persnr ) .